Add Measure function to the shim lock protocol
Expand the shim lock protocol to include a "Measure" functionality which can
be used to record hashes into the TPM of arbitrary buffers. The Measure
function will always measure the buffers directly (ie. not with Authenticode),
but for buffers that point to PE images the log will record that it was of a
PE image.

Signed-off-by: Tamas K Lengyel <lengyelt@ainfosec.com>